Extract TestCiUpload module
Slim down `lib/test_bot.rb`.
